Revert "Changed ref_count type to uint16 to see what happens"
This reverts commit 13dd2e7cad3b84e8ab5c5531e19988a190f9b503.

+++ src/rc_allocator.h
@@ -59,7 +59,7 @@ typedef struct rc_allocator {
    allocator               super;
    allocator_config       *cfg;
    buffer_handle           bh;
-   uint16                  *ref_count;
+   uint8                  *ref_count;
    uint64                  hand;
    io_handle              *io;
    rc_allocator_meta_page *meta_page;
